Add a plugin system to handle output.
This first commit implement the plugin system in passivedns and add
a new plugin that output in JSON format. You can define a plugin
with the -O argument in the command line.

This commit don't remove the JSON output (-j or -J arguments) of
passivedns.

You can test this plugin to output JSON message on stdout:
./src/passivedns -L - -l - -r dns-exemple.pcap -O
src/plugins/json/.libs/json.so -L - -l -